+++ App.js
@@ -1,11 +1,10 @@
 import React from 'react';
-import { StyleSheet, Text, View } from 'react-native';
-import Root from './Root';
+import Root from './components/Root';
 
 export default class App extends React.Component {
   render() {
     return (
-     <Root/>
+      <Root />
     );
   }
 }
